app: use flask to serve all components of GR
At the moment using flask to serve all the components of the game. This
includes the client page, stylesheets, javascript, assets and the API.
In the short term it may be possible to use this method with a public
webhost like pythonanywhere.com.  Later, as things grow, the non-API
components can be removed from the flask app and served by a proper
server (possibly requiring support of cross origin ajax requests).