client js: better random image funcs and change multipanel content
Improve the useability of the random image functions so that it is
easier to add a random image to an html tag for development.

Added a method to change the content of the multipanel when you click
one of the icons in the multipanel navbar.

Note: normally I'd break these two up but the order of the changes made
that non-trivial. I left them combined to save time.