makefile: reformat to use modular builds
Instead of using PHONY for every target, I'm creating the target files
using "date > target_file". Then in the clean targets the appropriate
target files should be also cleaned up as parts of the project get
cleaned. If you change a part of the project and want to rebuild you can
remove the specific target file and rebuild:

    Example:  rm client; make all && make drun