rollup.js: use umd format instead of common js
CommonJS was added to try and fix an issue I had with the javascript a
while ago. It seems that CommonJS and the "cjs" rollup format requires
additional javascript loader libraries which I do not want to use. UMD
seems to bundle the ES6 modules into a single browser acceptable bundle.