client: register browser events in gridrealm.js
Register all client events against the browser instead of using HTML
attributes like "onclick" to call the client javascript. This ensures
that rollup.js is able to figure out which code is used so that our
client code does not get dropped during the "treeshaking" process to
remove any dead code and reduce the browser payload.